Extract the user-scope, legacy, and claude-user-config policy codecs
Continue decomposing the setup surface by moving three self-contained
policy leaves out of cli/commands/setup.py into cli/integrations/:

- user_scope.py: _registered_project_keys + _user_scope_safe_to_clear,
  the registry-backed check that guards shared user-scope artifacts
  during a per-project teardown.
- legacy.py: the CLAUDE_MD_START/END markers and _remove_claude_md,
  which strip old injected Nauro blocks from CLAUDE.md.
- claude_user_config.py: _prune_redundant_user_scope_mcp, which drops a
  redundant user-scope HTTP nauro entry from ~/.claude.json.

Pure verbatim moves with no behavior change: the function bodies are
byte-identical to their prior definitions. setup.py imports each symbol
back for its own call sites, and the now-dead nauro.constants,
nauro.store.reader, and registry imports are removed. Test imports for
the moved symbols are retargeted to the new modules; assertions are
untouched.
