+++ packages/nauro/tests/test_setup_atomic_writes.py
@@ -33,10 +33,10 @@ from nauro.cli.commands.setup import (
     HOOK_TIMEOUT_SECONDS,
     _configure_mcp,
     _find_nauro_command,
-    _prune_redundant_user_scope_mcp,
     materialize_hooks_claude_code,
     materialize_hooks_codex,
 )
+from nauro.cli.integrations.claude_user_config import _prune_redundant_user_scope_mcp
 from nauro.store import _atomic
 
 pytestmark = pytest.mark.skipif(
