+++ packages/nauro/tests/test_setup_extended.py
@@ -15,8 +15,8 @@ from nauro.cli.commands.setup import (
     CHECK_HINT_LINE,
     _configure_codex,
     _configure_cursor_for_repo,
-    _prune_redundant_user_scope_mcp,
 )
+from nauro.cli.integrations.claude_user_config import _prune_redundant_user_scope_mcp
 from nauro.cli.integrations.skills import materialize_skills_cursor_for_repo
 from nauro.cli.main import app
 from nauro.store.registry import register_project_v2
