+++ packages/nauro/tests/test_setup_user_write_safety.py
@@ -24,9 +24,9 @@ import tomlkit
 from nauro.cli.commands.setup import (
     _configure_codex,
     _find_nauro_command,
-    _prune_redundant_user_scope_mcp,
 )
 from nauro.cli.integrations.agents import materialize_agents
+from nauro.cli.integrations.claude_user_config import _prune_redundant_user_scope_mcp
 from nauro.cli.integrations.skills import _materialize_skill_file, _remove_skill_file
 
 if sys.version_info >= (3, 11):
